refactor: separate invoice functionality and reorganize controllers
- Created dedicated InvoiceController for invoice-related functionality
- Updated invoice route to use {invoice} parameter with route model binding
- Added separate route for latest invoice (without parameter)
- Moved StudentController from Guardian namespace to main Controllers
- Removed invoice method from BillingController
- Updated tests to reflect new controller structure
